moved former ConsumerRecords to another package
