+++ clients/src/main/java/io/strimzi/common/records/consumer/http/ConsumerRecord.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.common.records.http.consumer;
+package io.strimzi.common.records.consumer.http;
 
 import java.util.Objects;
