+++ clients/src/main/java/io/strimzi/common/records/producer/http/ProducerRecord.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.common.records.http.producer;
+package io.strimzi.common.records.producer.http;
 
 import io.strimzi.test.tracing.HttpContext;
