+++ clients/src/main/java/io/strimzi/http/consumer/HttpConsumerClient.java
@@ -8,8 +8,8 @@ import io.grpc.netty.shaded.io.netty.handler.codec.http.HttpResponseStatus;
 import io.strimzi.common.ClientsInterface;
 import io.strimzi.configuration.ConfigurationConstants;
 import io.strimzi.configuration.http.HttpConsumerConfiguration;
-import io.strimzi.common.records.http.consumer.ConsumerRecord;
-import io.strimzi.common.records.http.consumer.ConsumerRecordUtils;
+import io.strimzi.common.records.consumer.http.ConsumerRecord;
+import io.strimzi.common.records.consumer.http.ConsumerRecordUtils;
 import io.strimzi.test.tracing.HttpContext;
 import io.strimzi.test.tracing.HttpHandle;
 import io.strimzi.test.tracing.TracingHandle;
