+++ clients/src/main/java/io/strimzi/http/producer/HttpProducerClient.java
@@ -8,9 +8,9 @@ import io.grpc.netty.shaded.io.netty.handler.codec.http.HttpResponseStatus;
 import io.strimzi.common.ClientsInterface;
 import io.strimzi.configuration.ConfigurationConstants;
 import io.strimzi.configuration.http.HttpProducerConfiguration;
-import io.strimzi.common.records.http.producer.OffsetRecordSent;
-import io.strimzi.common.records.http.producer.OffsetRecordSentUtils;
-import io.strimzi.common.records.http.producer.ProducerRecord;
+import io.strimzi.common.records.producer.http.OffsetRecordSent;
+import io.strimzi.common.records.producer.http.OffsetRecordSentUtils;
+import io.strimzi.common.records.producer.http.ProducerRecord;
 import io.strimzi.test.tracing.HttpContext;
 import io.strimzi.test.tracing.HttpHandle;
 import io.strimzi.test.tracing.TracingHandle;
