+++ clients/src/main/java/io/strimzi/kafka/KafkaConsumerClient.java
@@ -8,7 +8,7 @@ import io.strimzi.common.ClientsInterface;
 import io.strimzi.common.properties.KafkaProperties;
 import io.strimzi.configuration.ConfigurationConstants;
 import io.strimzi.configuration.kafka.KafkaConsumerConfiguration;
-import io.strimzi.models.KafkaConsumerRecord;
+import io.strimzi.common.records.consumer.kafka.KafkaConsumerRecord;
 import io.strimzi.test.tracing.TracingUtil;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
